requirements.txt
Flask==2.0.3
gtts==2.2.3
moviepy==1.0.3
Pillow==8.4.0
